% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{agrammon_options}
\alias{agrammon_options}
\title{Agrammon Options}
\usage{
agrammon_options(show = FALSE, ...)
}
\arguments{
\item{show}{logical. If \code{TRUE} an explanatory summary on the 
options that can be edited will be printed to the console. 
Defaults to \code{FALSE}.}

\item{\dots}{agrammon REST interface options}
}
\value{
a list of agrammon REST interface options
}
\description{
return a list of valid Agrammon REST interface options
}
\examples{
## examples here
agrammon_options(show = TRUE)
agrammon_options(language = 'de', print = c('LivestockNH3', 'LivestockTAN'))
}
