% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{check_and_validate}
\alias{check_and_validate}
\title{Check and Validate Agrammon Input}
\usage{
check_and_validate(dt, token = NULL)
}
\arguments{
\item{dt}{a \code{data.table} of Agrammon input data}
}
\value{
a list with entries 'input data', 'farm_id column', 'simulation column'
}
\description{
check and validate Agrammon input
}
