% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{check_report}
\alias{check_report}
\title{Check Report Argument}
\usage{
check_report(report)
}
\arguments{
\item{report}{string. Name of the report to generate.}
}
\value{
a vector of 'print-only' labels
}
\description{
check report argument and return corresponding 'print-only' labels
}
