% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_vars.r
\name{check_request}
\alias{check_request}
\title{Check Agrammon Return Value}
\usage{
check_request(req)
}
\arguments{
\item{req}{a list object returned by the call to Agrammon REST interface}
}
\value{
the Agrammon result as a string
}
\description{
check the returned content of the Agrammon REST call (POST request)
}
