% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{check_token}
\alias{check_token}
\title{Check Agrammon Access Token}
\usage{
check_token(token)
}
\arguments{
\item{token}{valid access token string of \code{NULL}}
}
\value{
a valid access token string
}
\description{
check and return a valid Agrammon access token
}
