% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_vars.r
\name{create_dataset}
\alias{create_dataset}
\title{create an input data set}
\usage{
create_dataset(
  ...,
  full_output = FALSE,
  data = NULL,
  dairy_cows = list(amount_summer = 1.5, amount_winter = 2.5, share_hay_summer = 100,
    share_maize_silage_summer = 0, share_maize_pellets_summer = 0,
    share_maize_silage_winter = 100, share_grass_silage_winter = 100,
    share_maize_pellets_winter = 0, share_potatoes_winter = 0, share_beets_winter = 0,
    milk_yield = 7500),
  pigs = list(energy_content = "defaults", crude_protein = "defaults",
    feeding_phase_1_crude_protein = 151, feeding_phase_2_crude_protein = 151,
    feeding_phase_3_crude_protein = 151),
  livestock_general = list(housing_type = "defaults", mitigation_housing_floor = "none",
    air_scrubber = "none", mitigation_housing_air = "none", free_range = "defaults",
    manure_removal_interval = "defaults", drinking_system = "defaults", yard_days =
    "defaults", yard_hours = "defaults", exercise_yard =
    "available_roughage_is_not_supplied_in_the_exercise_yard",
    floor_properties_exercise_yard = "defaults", grazing_days = "defaults", grazing_hours
    = "defaults"),
  storage_liquid = list(mixing_frequency = "7_to_12_times_per_year", cover_type =
    "solid_cover"),
  storage_solid = list(share_applied_direct_poultry_manure = 12, share_covered_basin =
    70, share_applied_direct_cattle_other_manure = 25, share_covered_basin_cattle_manure
    = 0, share_applied_direct_pig_manure = 0, share_covered_basin_pig_manure = 0),
  application_liquid = list(dilution_parts_water = 1, appl_rate = 25, fermented_slurry =
    0, appl_summer = 44, appl_autumn_winter_spring = 56, appl_evening = 20, appl_hotdays
    = "sometimes", share_splash_plate = 0, share_trailing_hose = 100, share_trailing_shoe
    = 0, share_shallow_injection = 0, share_deep_injection = 0),
  application_solid = list(incorp_lw1h = 0, incorp_lw4h = 0, incorp_lw8h = 0, incorp_lw1d
    = 0, incorp_lw3d = 0, incorp_gt3d = 0, incorp_none = 100, appl_summer = 30,
    appl_autumn_winter_spring = 70),
  mineral_fertiliser = list(),
  recycling_fertiliser = list(),
  token = NULL
)
}
\description{
create a valid input data set as R object (data.table)
}
\details{
create new data set
input for livestock animal categories: animal-category-label = list(animalcategory = ., ...)
  -> mandatory list entries: "animals" (number of animals)
input for slurry storage tanks: tank-label = list(...)
  -> mandatory list entries: "volume", "depth" and possibly "contains_cattle_manure"/"contains_pig_manure"
general input is provided as list without an argument name
run get_variable() to get an overview on all possible list entries
}
