% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_vars.r
\name{create_template}
\alias{create_template}
\title{create template in R}
\usage{
create_template(
  livestock = list(),
  storage = NULL,
  language = c("en", "de", "fr")[1],
  token = NULL
)
}
\arguments{
\item{x}{model dump as json}
}
\value{
a data.table containing all agrammon input variables
}
\description{
helper function to create template from model dump
}
\examples{
create_template(livestock = list(Equides = c('a' , 'b'), ponies_and_asses = c('c', 'd')))
x <- create_template(livestock = list(Equides = c('a' , 'b'), ponies_and_asses = c('c', 'd')), storage = c('tank1', 'tank2'))

create_template()
create_template(TRUE)
create_template(FALSE)
create_template('test1')
create_template(list('test2a'))
create_template(list('test2a', dairy_cows = 'b'))
create_template(list(dc = 'test2b', dairy_cows = 'b'))
create_template(list(dc = 'test2b', bc = 'a', dairy_cows = 'b'))
}
