% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_vars.r
\name{get_input_template}
\alias{get_input_template}
\title{get model dump}
\usage{
get_input_template(
  format = c("json", "csv", "text")[1],
  language = c("en", "de", "fr")[1],
  sort = c("model", "calculation")[1],
  token = NULL
)
}
\description{
helper function to get model dump via REST interface
}
\examples{
mj <- get_input_template()
}
