% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_vars.r
\name{get_tech_config}
\alias{get_tech_config}
\title{get technical config parameters}
\usage{
get_tech_config(file = "technical.cfg", token = NULL)
}
\value{
technical config parameters as data.table
}
\description{
function to get parameters from the technical.cfg file via REST interface
}
\examples{
# get standard technical config file
tech <- get_tech_config()
}
