% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_vars.r
\name{read_input_vars}
\alias{read_input_vars}
\title{process model dump}
\usage{
read_input_vars(x, language = "de", module = "")
}
\arguments{
\item{x}{model dump as json}
}
\description{
helper function to process model dump
}
\examples{
mj <- get_input_template()
x <- read_input_vars(mj)
}
