% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_token.r
\name{register_token}
\alias{register_token}
\title{Register token}
\usage{
register_token(token, permanent = TRUE)
}
\arguments{
\item{token}{Agrammon REST-API access token.}

\item{permanent}{store token permanently or only for current session. See Details.}
}
\description{
register an Agrammon REST interface access token for the current R session
or permanently
}
\details{
This function registers the \code{token} to access Agrammons REST interface in the 
current R session. 
If the argument \code{permanent} is \code{TRUE} (default), the 
provided \code{token} will be saved to the user Renviron file in the path 
\code{file.path(Sys.getenv("HOME"), ".Renviron")}.
The idea and the utilized code how to keep the token permanently is to a large part 
taken from the function \code{register_google} in the package \code{ggmap}.

Personal token: 

You can request a personal access token at support@agrammon.ch
}
\examples{
\dontrun{
  # register token permanently
  register_token(my_token)

  # register token for current session only
  register_token(my_token, permanent = FALSE)
}
}
\seealso{
[save_template()], [run_agrammon()]
}
