% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{report}
\alias{report}
\title{Extract Report}
\usage{
report(x, report = "summary")
}
\arguments{
\item{x}{results from a call to run_agrammon with argument \code{report = 'full'}}

\item{report}{string defining the content of the returned model result. 
Partial matching of argument.}
}
\value{
report based subset of results
}
\description{
extract report from Agrammon results
}
\examples{
## examples here
# path to example input data set
path_example <- system.file('extdata', 'example_data_set_3farms.csv',
  package = 'agrammon')
# run agrammon
res <- run_agrammon(path_example, report = 'full')
# return report on N flow
report(res, 'N')
}
