% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{run_agrammon}
\alias{run_agrammon}
\title{title}
\usage{
run_agrammon(
  input_file,
  report = c("summary", "detailed", "full", "NH3", "TAN", "N", "HAFL")[1],
  filter = c("total_only", "existing_categories", "all_categories")[1],
  language = c("en", "de", "fr")[1],
  token = NULL,
  ...
)
}
\arguments{
\item{token}{Agrammon REST-API access token. Can be provided as option entry 'agrammon.token'}
}
\value{
result from an Agrammon model run
}
\description{
description
}
\examples{
## examples here
run_agrammon('./tests/inputs-version6-rest.csv')
run_agrammon('./tests/inputs-version6-rest.csv', report = 'NH3', filter = 'ex')
}
