% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{run_model}
\alias{run_model}
\title{title}
\usage{
run_model(input, model_options = agrammon_options(), token = NULL)
}
\arguments{
\item{input}{either an input file name or an input data set as provided by create_dataset()}

\item{model_options}{options provided to the Agrammon model which have influence on model output results or}

\item{token}{Agrammon REST-API access token. Can be provided as option entry 'agrammon.token'}
}
\value{
result from an Agrammon model run
}
\description{
description
}
\examples{
run_model('./tests/inputs-version6-rest.csv')
run_model('./tests/inputs-version6-rest.csv', language = 'de')
}
