% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{save_excel}
\alias{save_excel}
\title{title}
\usage{
save_excel(x, file, wide_format = FALSE, asTable = TRUE)
}
\arguments{
\item{token}{Agrammon REST-API access token. Can be provided as option entry 'agrammon.token'}
}
\value{
result from an Agrammon model run
}
\description{
description
}
