% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_vars.r
\name{save_template}
\alias{save_template}
\title{save template to file}
\usage{
save_template(file, livestock = list(), storage = NULL, token = NULL)
}
\arguments{
\item{file}{file path}

\item{livestock}{a named list or \code{TRUE}. The default is an empty list. See Details.}

\item{storage}{a character vector providing storage instance names (slurry tanks)}

\item{token}{token which will be used to perform the REST call.}
}
\description{
save a csv file which can be used as template for model input
}
\details{
If \code{livestock} is an empty list (default), all available animal
categories including their parent classes are printed to the console.

If argument \code{livestock} is \code{TRUE}, entries for all available
parent classes will be written to the template file \code{file}.

If \code{livestock} is a named list, all defined instances in the list will be
written to the template file |code{file}. The list entry names must
correspond to a valid animal category or parent class, whereas the list entries
(character vectors) will provide the instance names of the corresponding animal
category (or parent class) (see Examples).

If \code{storage} is NULL, the first line of the template will contain a warning.

A valid \code{token} can be requested from support@agrammon.ch
}
\examples{
\dontrun{
  # register token permanently
  register_token(my_token)

  # save template to file
  save_template('model_template.csv', livestock = 
     list(Equides = c('Horses_1', 'Horses_2'), dairy_cows = 'DC'), 
     storage = 'Tank_1')
}
}
\seealso{
[register_token()] for registering your personal access token, [run_agrammon()]
for calling the Agrammon model with a valid input data set.
}
