% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{wide_output}
\alias{wide_output}
\title{title}
\usage{
wide_output(x)
}
\arguments{
\item{token}{Agrammon REST-API access token. Can be provided as option entry 'agrammon.token'}
}
\value{
result from an Agrammon model run
}
\description{
description
}
\examples{
## examples here
wide_output(res)
}
