% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_function.r
\name{run_agrammon}
\alias{run_agrammon}
\title{Run Agrammon}
\usage{
run_agrammon(
  input_file,
  report = getOption("agrammon.report", "full"),
  filter = getOption("agrammon.filter", "total_only"),
  language = getOption("agrammon.language", "en"),
  data.table = getOption("agrammon.datatable", TRUE),
  token = NULL,
  ...
)
}
\arguments{
\item{input_file}{path to CSV file containing model input data}

\item{report}{string defining the content of the returned model result. 
Check the table in the vignette to get more information on valid entries. 
Partial matching of argument.}

\item{filter}{string defining the detail with which the result is split up by animal category.
Valid entries are \code{'total_only'}, \code{'existing_categories'} or \code{'all_categories')}.
Partial matching of argument.}

\item{language}{language for the units in the model result. Valid entries are \code{'en'},
\code{'de'} or \code{'fr'}. Defaults to English ('en')}

\item{data.table}{logical. should the result be returned as \code{data.table} 
(\code{TRUE}; default) or \code{data.frame} (\code{FALSE}).}

\item{token}{Agrammon REST-API access token.}
}
\value{
result from an Agrammon model run
}
\description{
run the Agrammon model via its REST interface.
}
\examples{
## examples here
# path to example input data set
path_example <- system.file('extdata', 'example_data_set_3farms.csv',
  package = 'agrammon')
# return default report
run_agrammon(path_example)
# return report on NH3 losses and split up the result by all 
# animal categories that exist in the input data set
run_agrammon(path_example, report = 'NH3', filter = 'ex')
}
